Changes for leaching calculation
