+++ RUFAS/biophysical/manure/storage/open_lot_cbpb_calculator.py
@@ -29,6 +29,8 @@ DEFAULT_CARBON_FRACTION_AVAILABLE_IN_VSD = 0.5
 DEFAULT_CARBON_FRACTION_AVAILABLE_IN_VSND = 0.35
 """Default carbon content (percent by mass) of manure non-degradable volatile solids (unitless, [0, 1])."""
 
+LEACHING_COEFFICIENT: float = 0.035
+"""Leaching coefficient used in the calculation of nitrogen loss in a compost bedded pack barn (unitless)."""
 
 class OpenLotCbpbCalculator:
 
@@ -226,3 +228,30 @@ class OpenLotCbpbCalculator:
         return ManureConstants.EFFECTIVE_MICROBIAL_DECOMP_RATE * (
             math.pow(1.066, (temperature - 10)) - math.pow(1.21, (temperature - 50))
         )
+
+    @staticmethod
+    def calculate_nitrogen_loss_from_leaching(received_nitrogen: float) -> float:
+        """
+        Calculate the mass of nitrogen that leaches out of the manure-bedding mixture.
+
+        Parameters
+        ----------
+        received_nitrogen : float
+            The mass of nitrogen present in the manure excreted by animals (kg).
+
+        Returns
+        -------
+        float
+            The amount of nitrogen that leaches out of the mixture (kg).
+
+        Raises
+        ------
+        ValueError
+            If the daily nitrogen input is negative.
+
+        """
+
+        if received_nitrogen < 0.0:
+            raise ValueError(f"Daily nitrogen input mass must be non-negative: {received_nitrogen}")
+
+        return LEACHING_COEFFICIENT * received_nitrogen
